+++ leo-planer/src/main/java/at/htlleonding/leoplaner/dto/TimetableDTO.java
@@ -2,7 +2,7 @@ package at.htlleonding.leoplaner.dto;
 
 import java.util.List;
 
-public record TimetableDTO(int weeklyHours, List<ClassSubjectInstanceDTO> classSubjectInstances, int cost,
+public record TimetableDTO(int weeklyHours, List<ClassSubjectInstanceDTO> classSubjectInstances, long cost,
         double temperature) {
 
     public TimetableDTO {
